% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_pde_difusion.R
\name{actualizar_nutriente}
\alias{actualizar_nutriente}
\title{Un paso de integración explícita del campo de nutriente}
\usage{
actualizar_nutriente(N, estado, D, k, dt, dx = 1, N_half = 0.3)
}
\description{
Un paso de integración explícita del campo de nutriente
}
