% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_pde_difusion.R
\name{antibiotico_objetivo}
\alias{antibiotico_objetivo}
\title{Concentración de antibiótico "objetivo" en el paso actual: 0
antes de \code{paso_introduccion} (CAMBIO ABRUPTO), y la plantilla
espacial escalada por A_max y por una modulación temporal
después (heterogeneidad temporal: fluctuación).}
\usage{
antibiotico_objetivo(
  plantilla,
  paso,
  paso_introduccion,
  A_max = 1,
  modo_temporal = c("constante", "pulsos", "sinusoidal"),
  periodo = 40
)
}
\arguments{
\item{plantilla}{matriz espacial en \verb{[0,1]} (de \code{crear_plantilla_antibiotico})}

\item{paso}{paso de tiempo actual}

\item{paso_introduccion}{paso en el que ocurre el CAMBIO ABRUPTO}

\item{A_max}{concentración máxima post-introducción}

\item{modo_temporal}{"constante", "pulsos" o "sinusoidal"}

\item{periodo}{periodo (en pasos) de la fluctuación (pulsos/sinusoidal)}
}
\description{
Concentración de antibiótico "objetivo" en el paso actual: 0
antes de \code{paso_introduccion} (CAMBIO ABRUPTO), y la plantilla
espacial escalada por A_max y por una modulación temporal
después (heterogeneidad temporal: fluctuación).
}
