% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_analisis.R
\name{calcular_dimension_fractal}
\alias{calcular_dimension_fractal}
\title{Calcula la dimensión fractal del estado y devuelve también la tabla de conteos}
\usage{
calcular_dimension_fractal(estado, tamanos = c(2, 4, 5, 8, 10, 16, 20))
}
\arguments{
\item{estado}{matriz binaria de ocupación}

\item{tamanos}{vector de tamaños de caja a probar}
}
\value{
lista con \code{tabla}, \code{ajuste}, y \code{dimension}
}
\description{
Calcula la dimensión fractal del estado y devuelve también la tabla de conteos
}
