% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_analisis.R
\name{comparar_persistencia_empirica}
\alias{comparar_persistencia_empirica}
\title{Extrae tiempos de persistencia empíricos de una lista de réplicas HCA}
\usage{
comparar_persistencia_empirica(replicas)
}
\arguments{
\item{replicas}{lista devuelta por \code{simular_hca_replicas()}}
}
\value{
data.frame con replica, tiempo_persistencia, extinguio
}
\description{
Extrae tiempos de persistencia empíricos de una lista de réplicas HCA
}
