% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_campos.R
\name{crear_campo_nutriente}
\alias{crear_campo_nutriente}
\title{Crea el campo inicial de nutriente}
\usage{
crear_campo_nutriente(nx, ny, N0 = 1, ruido_sd = 0)
}
\arguments{
\item{nx, ny}{dimensiones de la retícula}

\item{N0}{concentración inicial homogénea de nutriente}

\item{ruido_sd}{ruido espacial opcional (0 = campo perfectamente homogéneo)}
}
\description{
Crea el campo inicial de nutriente
}
