% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_analisis.R
\name{dimension_fractal}
\alias{dimension_fractal}
\title{Dimensión fractal por conteo de cajas (box-counting)}
\usage{
dimension_fractal(estado, tamanos = c(2, 4, 5, 8, 10, 16, 20))
}
\arguments{
\item{estado}{matriz binaria de ocupación}

\item{tamanos}{vector de tamaños de caja a probar (potencias de 2 recomendadas)}
}
\value{
lista con el ajuste lineal log(N_cajas) ~ log(1/tamano) y la pendiente (dimensión)
}
\description{
Dimensión fractal por conteo de cajas (box-counting)
}
