% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_colonia.R
\name{fenotipo_z}
\alias{fenotipo_z}
\title{Fenotipo realizado z = g + e (solo definido donde estado==1)}
\usage{
fenotipo_z(g, e)
}
\description{
Fenotipo realizado z = g + e (solo definido donde estado==1)
}
