% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_ca_crecimiento.R
\name{fitness_estabilizador}
\alias{fitness_estabilizador}
\title{Fitness de selección estabilizadora individual (Caja 3, sin
integrar sobre la distribución poblacional: aquí se evalúa
directamente en el fenotipo REALIZADO z de cada agente)}
\usage{
fitness_estabilizador(theta_local, z, omega2)
}
\arguments{
\item{theta_local}{óptimo fenotípico local (matriz, misma forma que z)}

\item{z}{fenotipo realizado local (matriz)}

\item{omega2}{ancho^2 de la selección estabilizadora}
}
\description{
Fitness de selección estabilizadora individual (Caja 3, sin
integrar sobre la distribución poblacional: aquí se evalúa
directamente en el fenotipo REALIZADO z de cada agente)
}
