% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{hca_server}
\alias{hca_server}
\title{HCA app server function}
\usage{
hca_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}
}
\value{
A Shiny server function.
}
\description{
HCA app server function
}
