% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_genetica_cuantitativa.R
\name{lambda_DN_exponencial}
\alias{lambda_DN_exponencial}
\title{Caso particular de lambda_DN cuando la población silvestre declina
exponencialmente: N_w(t) = N0\emph{exp(-r}t), con tasa de mutación mu
constante por individuo. lambda_DN = mu*N0/r (ver Caja 4).}
\usage{
lambda_DN_exponencial(mu, N0, r)
}
\description{
Caso particular de lambda_DN cuando la población silvestre declina
exponencialmente: N_w(t) = N0\emph{exp(-r}t), con tasa de mutación mu
constante por individuo. lambda_DN = mu*N0/r (ver Caja 4).
}
