% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_campos.R
\name{laplaciano_periodico}
\alias{laplaciano_periodico}
\title{Laplaciano discreto 2D con condiciones de frontera PERIÓDICAS
(vectorizado, sin loops, usando desplazamiento circular de índices)}
\usage{
laplaciano_periodico(M, dx = 1)
}
\description{
Laplaciano discreto 2D con condiciones de frontera PERIÓDICAS
(vectorizado, sin loops, usando desplazamiento circular de índices)
}
