% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_genetica_cuantitativa.R
\name{Nt_analitico_qg}
\alias{Nt_analitico_qg}
\title{Trayectoria analítica de N_t tras un cambio ambiental abrupto
(fórmula de la sección "Abrupt environmental change" de la Caja 3,
evolución determinista, demografía densidad-independiente)}
\usage{
Nt_analitico_qg(t, N0, Wmax, omega2, sigma_g2, sigma_e2, d0)
}
\arguments{
\item{t}{vector de generaciones (t=0,1,2,...)}

\item{N0}{tamaño poblacional inicial (en el momento del shock)}

\item{Wmax}{aptitud máxima posible}

\item{omega2}{ancho^2 de la selección estabilizadora}

\item{sigma_g2}{varianza genética aditiva (constante, supuesto de la Caja 3)}

\item{sigma_e2}{varianza ambiental}

\item{d0}{desviación inicial del óptimo: d0 = theta1 - g_bar_0
(en el momento del shock, ANTES de que la población responda)}
}
\value{
vector N_t (misma longitud que \code{t})
}
\description{
Trayectoria analítica de N_t tras un cambio ambiental abrupto
(fórmula de la sección "Abrupt environmental change" de la Caja 3,
evolución determinista, demografía densidad-independiente)
}
