% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_genetica_cuantitativa.R
\name{prob_exceso_persistencia}
\alias{prob_exceso_persistencia}
\title{Probabilidad de EXCESO de persistencia por evolucionar vs. no
evolucionar (ecuación S9): diferencia entre sobrevivir-con-evolución
y sobrevivir-sin-evolución, en función del tiempo.}
\usage{
prob_exceso_persistencia(t, N0, m0, lambda_w, mu_w, lambda_m, mu_m)
}
\description{
Probabilidad de EXCESO de persistencia por evolucionar vs. no
evolucionar (ecuación S9): diferencia entre sobrevivir-con-evolución
y sobrevivir-sin-evolución, en función del tiempo.
}
