% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_genetica_cuantitativa.R
\name{prob_extincion_con_evolucion}
\alias{prob_extincion_con_evolucion}
\title{P(T_ext <= t | CON evolución): mezcla de m0 mutantes (lambda_m,mu_m)
y (N0-m0) silvestres (lambda_w,mu_w), sin nuevas mutaciones
(ecuación S5).}
\usage{
prob_extincion_con_evolucion(t, N0, m0, lambda_w, mu_w, lambda_m, mu_m)
}
\description{
P(T_ext <= t | CON evolución): mezcla de m0 mutantes (lambda_m,mu_m)
y (N0-m0) silvestres (lambda_w,mu_w), sin nuevas mutaciones
(ecuación S5).
}
