% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_genetica_cuantitativa.R
\name{prob_extincion_hasta_t}
\alias{prob_extincion_hasta_t}
\title{P(T_ext <= t) para un proceso de nacimiento-muerte puro con
tasa de nacimiento \code{lambda}, tasa de muerte \code{mu}, y n0 individuos
iniciales (ecuación S3/S6, forma general con n0 arbitrario).}
\usage{
prob_extincion_hasta_t(t, lambda, mu, n0)
}
\description{
P(T_ext <= t) para un proceso de nacimiento-muerte puro con
tasa de nacimiento \code{lambda}, tasa de muerte \code{mu}, y n0 individuos
iniciales (ecuación S3/S6, forma general con n0 arbitrario).
}
