% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_genetica_cuantitativa.R
\name{prob_rescate_asintotico}
\alias{prob_rescate_asintotico}
\title{Probabilidad ASINTÓTICA de rescate (t->Inf), ecuación S11:
solo depende de si el linaje mutante tiene lambda_m > mu_m.}
\usage{
prob_rescate_asintotico(lambda_m, mu_m, m0)
}
\description{
Probabilidad ASINTÓTICA de rescate (t->Inf), ecuación S11:
solo depende de si el linaje mutante tiene lambda_m > mu_m.
}
