% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_genetica_cuantitativa.R
\name{prob_rescate_de_novo}
\alias{prob_rescate_de_novo}
\title{Probabilidad de rescate vía mutación DE NOVO (ecuación B4.1),
asumiendo p_est constante en el tiempo.}
\usage{
prob_rescate_de_novo(lambda_DN, p_est)
}
\arguments{
\item{lambda_DN}{tasa esperada de linajes mutantes exitosos
(= integral de mu(t)*N_silvestre(t) en el tiempo; ver
\code{lambda_DN_exponencial()} para el caso de declive exponencial)}

\item{p_est}{probabilidad de establecimiento de un linaje mutante}
}
\description{
Probabilidad de rescate vía mutación DE NOVO (ecuación B4.1),
asumiendo p_est constante en el tiempo.
}
