% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_genetica_cuantitativa.R
\name{prob_rescate_total}
\alias{prob_rescate_total}
\title{Probabilidad TOTAL de rescate (ecuación B4.2): combina mutantes
preexistentes (varianza genética "standing", lambda_SGV,
aproximado Poisson) + mutantes de novo tras el shock.}
\usage{
prob_rescate_total(lambda_SGV, lambda_DN, p_est)
}
\description{
Probabilidad TOTAL de rescate (ecuación B4.2): combina mutantes
preexistentes (varianza genética "standing", lambda_SGV,
aproximado Poisson) + mutantes de novo tras el shock.
}
