% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_colonia.R
\name{resortear_ambiente}
\alias{resortear_ambiente}
\title{Resortea la componente ambiental \code{e} de TODOS los sitios ocupados
en el paso actual. Modela el supuesto de genética cuantitativa
de que e ~ N(0, sigma_e^2) i.i.d. en cada generación, independiente
del valor genético y del pasado del individuo.}
\usage{
resortear_ambiente(estado, e, sigma_e)
}
\arguments{
\item{estado}{matriz lógica de ocupación}

\item{e}{matriz de componente ambiental (se sobreescribe donde
estado == TRUE; se deja NA donde está vacío)}

\item{sigma_e}{sd de la componente ambiental}
}
\description{
Resortea la componente ambiental \code{e} de TODOS los sitios ocupados
en el paso actual. Modela el supuesto de genética cuantitativa
de que e ~ N(0, sigma_e^2) i.i.d. en cada generación, independiente
del valor genético y del pasado del individuo.
}
