% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_analisis.R
\name{sim_datos_shock}
\alias{sim_datos_shock}
\title{Extrae datos de choque del objeto \code{sim} devuelto por \code{simular_hca()}}
\usage{
sim_datos_shock(sim)
}
\arguments{
\item{sim}{lista devuelta por \code{simular_hca()}}
}
\value{
lista con paso de introducción, N0, g_bar0 y parámetros clave
}
\description{
Extrae datos de choque del objeto \code{sim} devuelto por \code{simular_hca()}
}
