% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tema_hca}
\alias{tema_hca}
\title{Tema ggplot compartido por todos los graficos de la app, para
mantener consistencia visual entre pestañas.}
\usage{
tema_hca()
}
\description{
Tema ggplot compartido por todos los graficos de la app, para
mantener consistencia visual entre pestañas.
}
