% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_ca_crecimiento.R
\name{vecinos_von_neumann}
\alias{vecinos_von_neumann}
\title{Índices de los 4 vecinos von Neumann con envoltura periódica (toroide)}
\usage{
vecinos_von_neumann(f, c, nx, ny)
}
\description{
Índices de los 4 vecinos von Neumann con envoltura periódica (toroide)
}
