% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_pde_difusion.R
\name{verificar_estabilidad_dt}
\alias{verificar_estabilidad_dt}
\title{Verifica la condición de estabilidad de von Neumann para FTCS 2D}
\usage{
verificar_estabilidad_dt(D, dt, dx = 1)
}
\description{
Verifica la condición de estabilidad de von Neumann para FTCS 2D
}
